% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust.R
\name{adjust}
\alias{adjust}
\title{Adjusts pooled p-values to Account for Multiple Hypothesis Testing}
\usage{
adjust(p.values)
}
\arguments{
\item{p.values}{All the pooled p-values for every biomarker}
}
\value{
Returns adjusted pooled p-values to account for multiple hypothesis testing.
}
\description{
The function, "adjust", reads in all of our pooled p-values.  We make a decision based on the number of p-values being sent to the function.  If this number is less than or equal to 10, we use function "p.adjust" on our p-values using method "bonferroni".  If this number is greater than 10, we use function "p.adjust" on our p-values using method "fdr".  Finally, we return the adjusted p-values to the user.
}
\examples{
pval <- c(.05, .5, .75)
adjusted.pval <- adjust(pval)
print(adjusted.pval)
}
