% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stouffer.pool.R
\name{stouffer.pool}
\alias{stouffer.pool}
\title{Stouffer's Method for Pooled P-values}
\usage{
stouffer.pool(p.values)
}
\arguments{
\item{p.values}{All the p-values for a specific biomarker}
}
\value{
Returns pooled p-value using Stouffer’s method.
}
\description{
The function, "stouffer.pool", reads in the p-values for a specific biomarker.  First, we find the total number of p-values sent to the function and store this information into a variable, 'k'.  We then create a variable "stat" to hold the value \eqn{\sum_{i=1}^{k} \phi^{-1}(p_i) / \sqrt{k}}.  We generate our pooled p-value utilizing normal distribution.  Finally, we return the pooled p-value for the biomarker.
}
\examples{
pval <- c(.05, .5, .75)
pooled.pval <- stouffer.pool(pval)
print(pooled.pval)
}
\references{
Chang, L., Lin, H., Sibille, E., & Tseng, G. C. (2013). Meta-analysis methods for combining multiple expression profiles: comparisons, statistical characterization and an application guideline. BMC Bioinformatics, 14(1).
https://doi.org/10.1186/1471-2105-14-368
}
