% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted.stouffer.pool.R
\name{weighted.stouffer.pool}
\alias{weighted.stouffer.pool}
\title{Weighted Stouffer's Method for Pooled P-values}
\usage{
weighted.stouffer.pool(p.values, n)
}
\arguments{
\item{p.values}{All the p-values for a specific biomarker}

\item{n}{A vector which contains the number of observations of the data set we used to generate each respective p-value}
}
\value{
Returns pooled p-value using Weighted Stouffer’s method.
}
\description{
The function, "weighted.stouffer.pool", reads in the p-values for a specific biomarker and the number of observations which generated each of these p-values, which we store into variable 'n'.  First, we find the total number of p-values sent to the function and store this information into a variable, 'k'.  Next, we set variable, "weight" equal to the \eqn{\sqrt{n}}.  I use method that pooled p-value equals \eqn{\phi \left(\dfrac{\sum_{i=1}^{k}\text{weight}_i \times \phi^{-1}(p_i)}{\sqrt{\sum_{i=1}^{k}\text{weight}_i^2}} \right)}.  Finally, we return the pooled p-value for the biomarker.
}
\examples{
pval <- c(.05, .5, .75)
n <- c(50, 100, 200)
pooled.pval <- weighted.stouffer.pool(pval, n)
print(pooled.pval)
}
\references{
Zaykin, D. V. (2011). Optimally weighted Z-test is a powerful method for combining probabilities in meta-analysis. Journal of Evolutionary Biology, 24(8), 1836–1841.
https://doi.org/10.1111/j.1420-9101.2011.02297.x
}
